feat: Improve period-selector impl
