+++ packages/frontend/src/pages/dashboard/types.ts
@@ -1,0 +1,4 @@
+export interface Period {
+  from: Date;
+  to: Date;
+}
